+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to this project will be documented in this file.
 
+## NEXT
+
+* Add build on node 14.x & issue templates
+
 ## 3.2.1 (2020-11-07)
 
 * Minor tweaks to Pushover notification events